Split rendering.h junk drawer: ObstacleChildren, ScreenTransform, mesh-render descriptors (refs #1194)
Next cut of the rendering.h SPLIT action item from #1194 (audit verdict:
"junk drawer: pass tags + size/layer wrappers + camera letterbox +
screen_to_virtual() helper + render descriptors + child-ownership").

## What moved

| Type                                  | New home                           | Concern                                      |
|---|---|---|
| ScreenTransform                       | components/camera_resources.h      | camera-system letterbox ctx singleton        |
| MeshType / ModelTransform / MeshChild | components/render_mesh.h (new)     | per-entity GPU mesh-render descriptors       |
| ObstacleChildren                      | components/obstacle.h              | obstacle-archetype child-entity ownership    |

camera_resources.h already houses the other camera-pipeline ctx singletons
(RenderTargets, ShapeMeshes, FloorParams), so ScreenTransform sits with its
authoritative producer/owner. render_mesh.h is a new header for the
universal entity-bound mesh-render descriptors that camera_system produces
and game_render_system consumes — these are emplaced on player, popup,
particle, obstacle, and mesh-child entities, so they don't belong in
obstacle.h. ObstacleChildren is obstacle-archetype-specific (literally named
that way) and now sits with the obstacle component table.

## Shim / back-compat

rendering.h becomes a thin shim that re-includes the new canonical headers,
matching the pattern from PR #1231 / #1233 / #1235. The wrapper-noise types
DrawSize / DrawLayer / Layer / ScreenPosition remain here pending the
wrapper deletion sweep (#1198).

game_render_system.cpp's only types from rendering.h were ModelTransform /
MeshType — those are all in render_mesh.h now, so the include switches to
the canonical header. Other consumers stay on the shim (they still need
DrawSize / Layer / ScreenPosition).

## Tests

tests/test_component_raylib_boundaries.cpp guarded the post-#1199 contract
that components/rendering.h uses raylib types directly (Matrix, Color,
"#include <raylib.h>") and never uses glm / Vec2f / Mat4f / TintColor
wrappers. Post-split that contract migrates to render_mesh.h (the canonical
home of Matrix and Color). The negative checks (no glm / Vec2f / Mat4f /
TintColor wrappers) are retained on both the shim and the new header.

## Build / tests

- cmake --build build       — zero warnings with -Wall -Wextra -Werror
- cmake --build build-no-unity — zero warnings (non-unity)
- ./build/shapeshifter_tests — All 2959 assertions in 902 test cases pass
  (+5 from the boundary test's new render_mesh.h checks)

## Out of scope

- game_state.h SPLIT (blocked on enum eradication ordering in #1202/#1204)
- DrawSize / DrawLayer / Layer / ScreenPosition wrapper deletion (tracked by #1198)
- Switching every shim consumer to canonical headers (deferred; matches
  PR #1231 / #1233 "patch direct consumers only" pattern)
